Feat: PrismaNet base class with hook for fetching intermediate activations
